pyproject.toml
[project]
name = "handwrite"
version = "0.1.0"
description = "AI Chinese handwriting generator"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0",
    "torchvision>=0.15",
    "Pillow>=10.0",
    "numpy>=1.24",
    "opencv-python>=4.8",
    "gradio>=4.0",
    "tqdm>=4.65",
    "pytesseract>=0.3.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "matplotlib>=3.7",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]


[tool.setuptools.package-data]
handwrite = [
    "assets/prototypes/*/manifest.json",
    "assets/prototypes/*/glyphs/*.png",
]
